tests/requirements.txt
PyYAML==5.1
seleniumbase==1.36.10
